refactor execution coordinator internals by responsibility
