+++ control-plane/aegisops_control_plane/execution_coordinator.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
-from dataclasses import replace
 from datetime import datetime, timezone
 import hashlib
 import json
-import logging
 from typing import Mapping, Protocol
 
 from .models import (
     ActionExecutionRecord,
     ActionRequestRecord,
-    ApprovalDecisionRecord,
     ReconciliationRecord,
 )
 
@@ -134,11 +131,21 @@ def _json_ready(value: object) -> object:
     return value
 
 
+from .execution_coordinator_action_requests import ReviewedActionRequestCoordinator
+from .execution_coordinator_delegation import ApprovedActionDelegationCoordinator
+from .execution_coordinator_reconciliation import (
+    ActionExecutionReconciliationCoordinator,
+)
+
+
 class ExecutionCoordinator:
-    """Owns approved action request, delegation, and reconciliation flows."""
+    """Owns stable coordination entrypoints over extracted internals."""
 
     def __init__(self, service: ExecutionCoordinatorServiceDependencies) -> None:
         self._service = service
+        self._action_requests = ReviewedActionRequestCoordinator(service)
+        self._delegation = ApprovedActionDelegationCoordinator(service)
+        self._reconciliation = ActionExecutionReconciliationCoordinator(service)
 
     def create_reviewed_action_request_from_advisory(
         self,
@@ -152,165 +159,16 @@ class ExecutionCoordinator:
         expires_at: datetime,
         action_request_id: str | None = None,
     ) -> ActionRequestRecord:
-        requester_identity = self._service._require_non_empty_string(
-            requester_identity,
-            "requester_identity",
+        return self._action_requests.create_reviewed_action_request_from_advisory(
+            record_family=record_family,
+            record_id=record_id,
+            requester_identity=requester_identity,
+            recipient_identity=recipient_identity,
+            message_intent=message_intent,
+            escalation_reason=escalation_reason,
+            expires_at=expires_at,
+            action_request_id=action_request_id,
         )
-        recipient_identity = self._service._require_non_empty_string(
-            recipient_identity,
-            "recipient_identity",
-        )
-        message_intent = self._service._require_non_empty_string(
-            message_intent,
-            "message_intent",
-        )
-        escalation_reason = self._service._require_non_empty_string(
-            escalation_reason,
-            "escalation_reason",
-        )
-        expires_at = self._service._require_aware_datetime(expires_at, "expires_at")
-
-        with self._service._store.transaction():
-            context_snapshot = self._service.inspect_assistant_context(
-                record_family,
-                record_id,
-            )
-            self._service._require_reviewed_case_scoped_advisory_read(context_snapshot)
-            recommendation_draft = self._service.render_recommendation_draft(
-                record_family,
-                record_id,
-            )
-            if recommendation_draft.recommendation_draft.get("status") != "ready":
-                raise ValueError(
-                    "reviewed advisory context is not ready for approval-bound action requests"
-                )
-
-            recommendation_id = self._service._require_single_recommendation_binding(
-                record_family=record_family,
-                record_id=record_id,
-                linked_recommendation_ids=recommendation_draft.linked_recommendation_ids,
-            )
-            case_id = self._service._require_single_linked_case_id(
-                recommendation_draft.linked_case_ids
-            )
-            case = self._service._require_reviewed_operator_case(case_id)
-            if expires_at <= datetime.now(timezone.utc):
-                raise ValueError("expires_at must be in the future")
-
-            requested_payload = {
-                "action_type": "notify_identity_owner",
-                "recipient_identity": recipient_identity,
-                "message_intent": message_intent,
-                "escalation_reason": escalation_reason,
-                "source_record_family": record_family,
-                "source_record_id": record_id,
-                "recommendation_id": recommendation_id,
-                "case_id": case.case_id,
-                "alert_id": case.alert_id,
-                "finding_id": case.finding_id,
-                "linked_evidence_ids": recommendation_draft.linked_evidence_ids,
-            }
-            target_scope = {
-                "record_family": record_family,
-                "record_id": record_id,
-                "case_id": case.case_id,
-                "alert_id": case.alert_id,
-                "finding_id": case.finding_id,
-                "recipient_identity": recipient_identity,
-            }
-            payload_hash = _approved_payload_binding_hash(
-                target_scope=target_scope,
-                approved_payload=requested_payload,
-                execution_surface_type="automation_substrate",
-                execution_surface_id="shuffle",
-            )
-            requested_at = datetime.now(timezone.utc)
-            if expires_at <= requested_at:
-                raise ValueError("expires_at must be after requested_at")
-
-            idempotency_material = json.dumps(
-                _json_ready(
-                    {
-                        "payload_hash": payload_hash,
-                        "record_family": record_family,
-                        "record_id": record_id,
-                        "requester_identity": requester_identity,
-                        "expires_at": expires_at,
-                    }
-                ),
-                sort_keys=True,
-                separators=(",", ":"),
-            )
-            idempotency_key = (
-                "notify-identity-owner:"
-                + hashlib.sha256(idempotency_material.encode("utf-8")).hexdigest()
-            )
-            for existing in self._service._store.list(ActionRequestRecord):
-                if existing.idempotency_key == idempotency_key:
-                    self._service._emit_structured_event(
-                        logging.INFO,
-                        "action_request_reused",
-                        action_request_id=existing.action_request_id,
-                        action_type=existing.requested_payload.get("action_type"),
-                        lifecycle_state=existing.lifecycle_state,
-                        case_id=existing.case_id,
-                    )
-                    return existing
-
-            normalized_action_request_id = self._service._resolve_new_record_identifier(
-                ActionRequestRecord,
-                action_request_id,
-                "action_request_id",
-                "action-request",
-            )
-            created_request = self._service.persist_record(
-                ActionRequestRecord(
-                    action_request_id=normalized_action_request_id,
-                    approval_decision_id=None,
-                    case_id=case.case_id,
-                    alert_id=case.alert_id,
-                    finding_id=case.finding_id,
-                    idempotency_key=idempotency_key,
-                    target_scope=target_scope,
-                    payload_hash=payload_hash,
-                    requested_at=requested_at,
-                    expires_at=expires_at,
-                    lifecycle_state="pending_approval",
-                    requester_identity=requester_identity,
-                    requested_payload=requested_payload,
-                    policy_basis={
-                        "severity": "low",
-                        "target_scope": "single_identity",
-                        "action_reversibility": "reversible",
-                        "asset_criticality": "standard",
-                        "identity_criticality": "standard",
-                        "blast_radius": "single_target",
-                        "execution_constraint": "routine_allowed",
-                    },
-                    policy_evaluation={
-                        "approval_requirement": "human_required",
-                        "approval_requirement_override": "human_required",
-                        "routing_target": "approval",
-                        "execution_surface_type": "automation_substrate",
-                        "execution_surface_id": "shuffle",
-                    },
-                ),
-                transitioned_at=requested_at,
-            )
-        self._service._emit_structured_event(
-            logging.INFO,
-            "action_request_created",
-            action_request_id=created_request.action_request_id,
-            action_type=created_request.requested_payload.get("action_type"),
-            requester_identity=created_request.requester_identity,
-            case_id=created_request.case_id,
-            alert_id=created_request.alert_id,
-            lifecycle_state=created_request.lifecycle_state,
-            expires_at=created_request.expires_at.isoformat()
-            if created_request.expires_at is not None
-            else None,
-        )
-        return created_request
 
     def delegate_approved_action_to_shuffle(
         self,
@@ -321,22 +179,12 @@ class ExecutionCoordinator:
         delegation_issuer: str,
         evidence_ids: tuple[str, ...] = (),
     ) -> ActionExecutionRecord:
-        return self._delegate_approved_action(
+        return self._delegation.delegate_approved_action_to_shuffle(
             action_request_id=action_request_id,
             approved_payload=approved_payload,
             delegated_at=delegated_at,
             delegation_issuer=delegation_issuer,
             evidence_ids=evidence_ids,
-            execution_surface_type="automation_substrate",
-            execution_surface_id="shuffle",
-            invalid_execution_surface_type_message=(
-                "approved action request is not delegated through the automation "
-                "substrate path"
-            ),
-            invalid_execution_surface_id_message=(
-                "approved action request is not routed to the reviewed shuffle adapter"
-            ),
-            delegation_label="shuffle",
         )
 
     def delegate_approved_action_to_isolated_executor(
@@ -348,22 +196,12 @@ class ExecutionCoordinator:
         delegation_issuer: str,
         evidence_ids: tuple[str, ...] = (),
     ) -> ActionExecutionRecord:
-        return self._delegate_approved_action(
+        return self._delegation.delegate_approved_action_to_isolated_executor(
             action_request_id=action_request_id,
             approved_payload=approved_payload,
             delegated_at=delegated_at,
             delegation_issuer=delegation_issuer,
             evidence_ids=evidence_ids,
-            execution_surface_type="executor",
-            execution_surface_id="isolated-executor",
-            invalid_execution_surface_type_message=(
-                "approved action request is not delegated through the isolated "
-                "executor path"
-            ),
-            invalid_execution_surface_id_message=(
-                "approved action request is not routed to the reviewed isolated executor"
-            ),
-            delegation_label="isolated executor",
         )
 
     def reconcile_action_execution(
@@ -376,999 +214,21 @@ class ExecutionCoordinator:
         compared_at: datetime,
         stale_after: datetime,
     ) -> ReconciliationRecord:
-        compared_at = self._service._require_aware_datetime(compared_at, "compared_at")
-        stale_after = self._service._require_aware_datetime(stale_after, "stale_after")
-        execution_surface_type = self._service._require_non_empty_string(
-            execution_surface_type,
-            "execution_surface_type",
-        )
-        execution_surface_id = self._service._require_non_empty_string(
-            execution_surface_id,
-            "execution_surface_id",
-        )
-        action_request = self._service._store.get(ActionRequestRecord, action_request_id)
-        if action_request is None:
-            raise LookupError(f"Missing action request {action_request_id!r}")
-        if action_request.lifecycle_state != "approved":
-            raise ValueError(
-                f"Action request {action_request_id!r} is not approved "
-                f"(state={action_request.lifecycle_state!r})"
-            )
-
-        require_binding_identifiers = (
-            execution_surface_type == "automation_substrate"
-            and execution_surface_id == "shuffle"
-        )
-        require_coordination_receipt_identifiers = (
-            require_binding_identifiers
-            and action_request.requested_payload.get("action_type")
-            == "create_tracking_ticket"
-        )
-        normalized_executions = self._normalize_observed_executions(
-            observed_executions,
-            require_binding_identifiers=require_binding_identifiers,
-            require_coordination_receipt_identifiers=(
-                require_coordination_receipt_identifiers
-            ),
-        )
-        linked_execution_run_ids = tuple(
-            execution["execution_run_id"] for execution in normalized_executions
-        )
-        unique_execution_run_ids = tuple(dict.fromkeys(linked_execution_run_ids))
-        latest_execution = normalized_executions[-1] if normalized_executions else None
-        authoritative_execution = self._find_authoritative_action_execution(
-            action_request_id=action_request.action_request_id,
+        return self._reconciliation.reconcile_action_execution(
+            action_request_id=action_request_id,
             execution_surface_type=execution_surface_type,
             execution_surface_id=execution_surface_id,
-            execution_run_id=(
-                None if latest_execution is None else latest_execution["execution_run_id"]
-            ),
-            idempotency_key=(
-                action_request.idempotency_key
-                if latest_execution is None
-                else latest_execution["idempotency_key"]
-            ),
+            observed_executions=observed_executions,
+            compared_at=compared_at,
+            stale_after=stale_after,
         )
-
-        subject_linkage: dict[str, object] = {
-            "action_request_ids": (action_request.action_request_id,),
-            "execution_surface_types": (execution_surface_type,),
-            "execution_surface_ids": (execution_surface_id,),
-        }
-        if action_request.approval_decision_id is not None:
-            subject_linkage["approval_decision_ids"] = (
-                action_request.approval_decision_id,
-            )
-        if action_request.alert_id is not None:
-            subject_linkage["alert_ids"] = (action_request.alert_id,)
-        if action_request.case_id is not None:
-            subject_linkage["case_ids"] = (action_request.case_id,)
-        if action_request.finding_id is not None:
-            subject_linkage["finding_ids"] = (action_request.finding_id,)
-        if authoritative_execution is not None:
-            subject_linkage["action_execution_ids"] = (
-                authoritative_execution.action_execution_id,
-            )
-            subject_linkage["delegation_ids"] = (
-                authoritative_execution.delegation_id,
-            )
-            evidence_ids = self._service._merge_linked_ids(
-                authoritative_execution.provenance.get("evidence_ids"),
-                None,
-            )
-            if evidence_ids:
-                subject_linkage["evidence_ids"] = evidence_ids
-
-        ingest_disposition: str
-        lifecycle_state: str
-        mismatch_summary: str
-        execution_run_id: str | None = None
-        last_seen_at = action_request.requested_at
-
-        if latest_execution is None:
-            ingest_disposition = "missing"
-            lifecycle_state = "pending"
-            mismatch_summary = (
-                "missing downstream execution for approved action request correlation"
-            )
-        else:
-            execution_run_id = latest_execution["execution_run_id"]
-            last_seen_at = latest_execution["observed_at"]
-            observed_execution_surface_id = latest_execution["execution_surface_id"]
-            observed_idempotency_key = latest_execution["idempotency_key"]
-            observed_approval_decision_id = latest_execution.get("approval_decision_id")
-            observed_delegation_id = latest_execution.get("delegation_id")
-            observed_payload_hash = latest_execution.get("payload_hash")
-            observed_coordination_reference_id = latest_execution.get(
-                "coordination_reference_id"
-            )
-            observed_coordination_target_type = latest_execution.get(
-                "coordination_target_type"
-            )
-            observed_external_receipt_id = latest_execution.get("external_receipt_id")
-            observed_coordination_target_id = latest_execution.get(
-                "coordination_target_id"
-            )
-            observed_ticket_reference_url = latest_execution.get("ticket_reference_url")
-            expected_execution_run_id = (
-                None
-                if authoritative_execution is None
-                else authoritative_execution.execution_run_id
-            )
-            if last_seen_at < stale_after and compared_at >= stale_after:
-                ingest_disposition = "stale"
-                lifecycle_state = "stale"
-                mismatch_summary = "stale downstream execution observation requires refresh"
-            elif len(unique_execution_run_ids) > 1:
-                ingest_disposition = "duplicate"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "duplicate downstream executions observed for one approved request"
-                )
-            elif (
-                observed_execution_surface_id != execution_surface_id
-                or observed_idempotency_key != action_request.idempotency_key
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "execution surface/idempotency mismatch between approved request and observed execution"
-                )
-            elif (
-                expected_execution_run_id is not None
-                and execution_run_id != expected_execution_run_id
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "execution run identity mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            elif (
-                authoritative_execution is not None
-                and authoritative_execution.execution_surface_type == "automation_substrate"
-                and authoritative_execution.execution_surface_id == "shuffle"
-                and (
-                    observed_approval_decision_id
-                    != authoritative_execution.approval_decision_id
-                    or observed_delegation_id != authoritative_execution.delegation_id
-                    or observed_payload_hash != authoritative_execution.payload_hash
-                )
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "approved binding mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            elif (
-                require_coordination_receipt_identifiers
-                and authoritative_execution is None
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "coordination receipt mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            elif (
-                authoritative_execution is not None
-                and authoritative_execution.approved_payload.get("action_type")
-                == "create_tracking_ticket"
-                and authoritative_execution.execution_surface_type
-                == "automation_substrate"
-                and authoritative_execution.execution_surface_id == "shuffle"
-                and (
-                    observed_coordination_reference_id
-                    != authoritative_execution.provenance.get(
-                        "downstream_binding",
-                        {},
-                    ).get("coordination_reference_id")
-                    or observed_coordination_target_type
-                    != authoritative_execution.provenance.get(
-                        "downstream_binding",
-                        {},
-                    ).get("coordination_target_type")
-                    or observed_external_receipt_id
-                    != authoritative_execution.provenance.get(
-                        "downstream_binding",
-                        {},
-                    ).get("external_receipt_id")
-                    or observed_coordination_target_id
-                    != authoritative_execution.provenance.get(
-                        "downstream_binding",
-                        {},
-                    ).get("coordination_target_id")
-                    or observed_ticket_reference_url
-                    != authoritative_execution.provenance.get(
-                        "downstream_binding",
-                        {},
-                    ).get("ticket_reference_url")
-                )
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "coordination receipt mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            else:
-                ingest_disposition = "matched"
-                lifecycle_state = "matched"
-                mismatch_summary = (
-                    "matched approved action request to reviewed execution run"
-                )
-
-        with self._service._store.transaction():
-            if authoritative_execution is not None:
-                stored_execution = self._service._store.get(
-                    ActionExecutionRecord,
-                    authoritative_execution.action_execution_id,
-                )
-                if stored_execution is None:
-                    raise LookupError(
-                        "missing authoritative action execution during reconciliation"
-                    )
-                authoritative_execution = stored_execution
-
-            if (
-                authoritative_execution is not None
-                and latest_execution is not None
-                and ingest_disposition == "matched"
-            ):
-                reconciled_lifecycle_state = self._action_execution_lifecycle_from_status(
-                    latest_execution.get("status"),
-                    authoritative_execution.lifecycle_state,
-                )
-                if reconciled_lifecycle_state != authoritative_execution.lifecycle_state:
-                    authoritative_execution = self._service.persist_record(
-                        replace(
-                            authoritative_execution,
-                            lifecycle_state=reconciled_lifecycle_state,
-                        ),
-                        transitioned_at=latest_execution["observed_at"],
-                    )
-            if (
-                authoritative_execution is not None
-                and authoritative_execution.approved_payload.get("action_type")
-                == "create_tracking_ticket"
-            ):
-                downstream_binding = authoritative_execution.provenance.get(
-                    "downstream_binding",
-                    {},
-                )
-                if isinstance(downstream_binding, Mapping):
-                    if isinstance(downstream_binding.get("coordination_reference_id"), str):
-                        subject_linkage["coordination_reference_ids"] = (
-                            downstream_binding["coordination_reference_id"],
-                        )
-                    if isinstance(downstream_binding.get("coordination_target_type"), str):
-                        subject_linkage["coordination_target_types"] = (
-                            downstream_binding["coordination_target_type"],
-                        )
-                    if isinstance(downstream_binding.get("external_receipt_id"), str):
-                        subject_linkage["external_receipt_ids"] = (
-                            downstream_binding["external_receipt_id"],
-                        )
-                    if isinstance(downstream_binding.get("coordination_target_id"), str):
-                        subject_linkage["coordination_target_ids"] = (
-                            downstream_binding["coordination_target_id"],
-                        )
-                    if isinstance(downstream_binding.get("ticket_reference_url"), str):
-                        subject_linkage["ticket_reference_urls"] = (
-                            downstream_binding["ticket_reference_url"],
-                        )
-
-            reconciliation = self._service.persist_record(
-                ReconciliationRecord(
-                    reconciliation_id=self._service._next_identifier("reconciliation"),
-                    subject_linkage=subject_linkage,
-                    alert_id=action_request.alert_id,
-                    finding_id=action_request.finding_id,
-                    analytic_signal_id=None,
-                    execution_run_id=execution_run_id,
-                    linked_execution_run_ids=linked_execution_run_ids,
-                    correlation_key=self._build_action_execution_reconciliation_key(
-                        action_request=action_request,
-                        execution_surface_type=execution_surface_type,
-                        execution_surface_id=execution_surface_id,
-                        authoritative_execution=authoritative_execution,
-                    ),
-                    first_seen_at=action_request.requested_at,
-                    last_seen_at=last_seen_at,
-                    ingest_disposition=ingest_disposition,
-                    mismatch_summary=mismatch_summary,
-                    compared_at=compared_at,
-                    lifecycle_state=lifecycle_state,
-                ),
-                transitioned_at=compared_at,
-            )
-        self._service._emit_structured_event(
-            logging.INFO,
-            "action_execution_reconciled",
-            reconciliation_id=reconciliation.reconciliation_id,
-            action_request_id=action_request.action_request_id,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-            ingest_disposition=reconciliation.ingest_disposition,
-            lifecycle_state=reconciliation.lifecycle_state,
-            execution_run_id=reconciliation.execution_run_id,
-        )
-        return reconciliation
-
-    def _delegate_approved_action(
-        self,
-        *,
-        action_request_id: str,
-        approved_payload: Mapping[str, object],
-        delegated_at: datetime,
-        delegation_issuer: str,
-        evidence_ids: tuple[str, ...],
-        execution_surface_type: str,
-        execution_surface_id: str,
-        invalid_execution_surface_type_message: str,
-        invalid_execution_surface_id_message: str,
-        delegation_label: str,
-    ) -> ActionExecutionRecord:
-        delegated_at = self._service._require_aware_datetime(delegated_at, "delegated_at")
-        action_request_id = self._service._require_non_empty_string(
-            action_request_id,
-            "action_request_id",
-        )
-        delegation_issuer = self._service._require_non_empty_string(
-            delegation_issuer,
-            "delegation_issuer",
-        )
-        normalized_payload = self._service._require_mapping(
-            approved_payload,
-            "approved_payload",
-        )
-        adapter = (
-            self._service._shuffle
-            if execution_surface_id == "shuffle"
-            else self._service._isolated_executor
-        )
-        predispatch_execution: ActionExecutionRecord
-        action_request: ActionRequestRecord
-        approval_decision: ApprovalDecisionRecord
-        approval_decision_id: str
-        with self._service._store.transaction():
-            action_request, approval_decision = self._load_approved_delegation_context(
-                action_request_id=action_request_id,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-                execution_surface_type=execution_surface_type,
-                execution_surface_id=execution_surface_id,
-                invalid_execution_surface_type_message=(
-                    invalid_execution_surface_type_message
-                ),
-                invalid_execution_surface_id_message=(
-                    invalid_execution_surface_id_message
-                ),
-                delegation_label=delegation_label,
-            )
-            approval_decision_id = approval_decision.approval_decision_id
-            for existing in self._service._store.list(ActionExecutionRecord):
-                if (
-                    existing.action_request_id == action_request.action_request_id
-                    and existing.execution_surface_type == execution_surface_type
-                    and existing.execution_surface_id == execution_surface_id
-                    and existing.idempotency_key == action_request.idempotency_key
-                ):
-                    if existing.lifecycle_state == "dispatching":
-                        raise RuntimeError(
-                            "approved action delegation is already dispatching"
-                        )
-                    return existing
-            if execution_surface_id == "shuffle":
-                self._require_reviewed_shuffle_payload(normalized_payload)
-
-            delegation_id = self._service._next_identifier("delegation")
-            predispatch_execution = self._service.persist_record(
-                ActionExecutionRecord(
-                    action_execution_id=self._service._next_identifier("action-execution"),
-                    action_request_id=action_request.action_request_id,
-                    approval_decision_id=approval_decision_id,
-                    delegation_id=delegation_id,
-                    execution_surface_type=execution_surface_type,
-                    execution_surface_id=execution_surface_id,
-                    execution_run_id=self._pending_dispatch_execution_run_id(
-                        delegation_id
-                    ),
-                    idempotency_key=action_request.idempotency_key,
-                    target_scope=action_request.target_scope,
-                    approved_payload=normalized_payload,
-                    payload_hash=action_request.payload_hash,
-                    delegated_at=delegated_at,
-                    expires_at=action_request.expires_at,
-                    provenance={
-                        "delegation_issuer": delegation_issuer,
-                        "evidence_ids": evidence_ids,
-                    },
-                    lifecycle_state="dispatching",
-                ),
-                transitioned_at=delegated_at,
-            )
-        receipt = None
-        execution_run_id: str
-        try:
-            receipt = adapter.dispatch_approved_action(
-                delegation_id=delegation_id,
-                action_request_id=action_request.action_request_id,
-                approval_decision_id=approval_decision_id,
-                payload_hash=action_request.payload_hash,
-                idempotency_key=action_request.idempotency_key,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-            )
-            self._require_exact_adapter_receipt_binding(
-                receipt=receipt,
-                action_request=action_request,
-                approval_decision_id=approval_decision_id,
-                delegation_id=delegation_id,
-                execution_surface_type=execution_surface_type,
-                execution_surface_id=execution_surface_id,
-            )
-            execution_run_id = self._require_receipt_string_attribute(
-                receipt,
-                "execution_run_id",
-            )
-        except Exception as exc:
-            self._mark_dispatch_failure(
-                execution=predispatch_execution,
-                error=exc,
-                receipt=receipt,
-            )
-            raise
-
-        try:
-            with self._service._store.transaction():
-                stored_execution = self._service._store.get(
-                    ActionExecutionRecord,
-                    predispatch_execution.action_execution_id,
-                )
-                if stored_execution is None:
-                    raise LookupError(
-                        "missing pre-dispatch action execution record during delegation finalization"
-                    )
-                execution = self._service.persist_record(
-                    replace(
-                        stored_execution,
-                        execution_run_id=execution_run_id,
-                        provenance=self._finalized_execution_provenance(
-                            execution=stored_execution,
-                            receipt=receipt,
-                            execution_surface_id=execution_surface_id,
-                        ),
-                        lifecycle_state="queued",
-                    ),
-                    transitioned_at=delegated_at,
-                )
-        except Exception as exc:
-            self._mark_dispatch_failure(
-                execution=predispatch_execution,
-                error=exc,
-                receipt=receipt,
-            )
-            raise
-        self._service._emit_action_execution_delegated_event(execution)
-        return execution
-
-    @staticmethod
-    def _pending_dispatch_execution_run_id(delegation_id: str) -> str:
-        return f"pending-dispatch-{delegation_id}"
-
-    def _require_reviewed_shuffle_payload(
-        self,
-        approved_payload: Mapping[str, object],
-    ) -> None:
-        action_type = approved_payload.get("action_type")
-        if action_type == "notify_identity_owner":
-            for field_name in (
-                "recipient_identity",
-                "message_intent",
-                "escalation_reason",
-            ):
-                value = approved_payload.get(field_name)
-                if not isinstance(value, str) or value.strip() == "":
-                    raise ValueError(
-                        "approved action is outside the reviewed Phase 20 Shuffle delegation scope"
-                    )
-            return
-        if action_type == "create_tracking_ticket":
-            for field_name in (
-                "case_id",
-                "coordination_reference_id",
-                "coordination_target_type",
-                "ticket_title",
-                "ticket_description",
-            ):
-                value = approved_payload.get(field_name)
-                if not isinstance(value, str) or value.strip() == "":
-                    raise ValueError(
-                        "approved action is outside the reviewed Phase 26 Shuffle delegation scope"
-                    )
-            coordination_target_type = approved_payload.get("coordination_target_type")
-            if (
-                not isinstance(coordination_target_type, str)
-                or coordination_target_type
-                not in _PHASE26_REVIEWED_COORDINATION_TARGET_TYPES
-            ):
-                raise ValueError(
-                    "approved action is outside the reviewed Phase 26 Shuffle delegation scope"
-                )
-            return
-        raise ValueError(
-            "approved action is outside the reviewed Phase 20 Shuffle delegation scope"
-        )
-
-    def _require_exact_adapter_receipt_binding(
-        self,
-        *,
-        receipt: object,
-        action_request: ActionRequestRecord,
-        approval_decision_id: str,
-        delegation_id: str,
-        execution_surface_type: str,
-        execution_surface_id: str,
-    ) -> None:
-        if getattr(receipt, "execution_surface_type", None) != execution_surface_type:
-            raise ValueError("adapter receipt does not match approved execution surface")
-        if getattr(receipt, "execution_surface_id", None) != execution_surface_id:
-            raise ValueError("adapter receipt does not match approved execution surface")
-        if execution_surface_id == "shuffle" and any(
-            (
-                getattr(receipt, "approval_decision_id", None) != approval_decision_id,
-                getattr(receipt, "delegation_id", None) != delegation_id,
-                getattr(receipt, "payload_hash", None) != action_request.payload_hash,
-            )
-        ):
-            raise ValueError(
-                "shuffle receipt does not match approved delegation binding"
-            )
-        if (
-            execution_surface_id == "shuffle"
-            and action_request.requested_payload.get("action_type")
-            == "create_tracking_ticket"
-        ):
-            approved_coordination_reference_id = self._service._require_non_empty_string(
-                action_request.target_scope.get("coordination_reference_id"),
-                "action_request.target_scope.coordination_reference_id",
-            )
-            approved_coordination_target_type = self._service._require_non_empty_string(
-                action_request.target_scope.get("coordination_target_type"),
-                "action_request.target_scope.coordination_target_type",
-            )
-            if any(
-                (
-                    self._require_receipt_string_attribute(
-                        receipt,
-                        "coordination_reference_id",
-                    )
-                    != approved_coordination_reference_id,
-                    self._require_receipt_string_attribute(
-                        receipt,
-                        "coordination_target_type",
-                    )
-                    != approved_coordination_target_type,
-                )
-            ):
-                raise ValueError(
-                    "shuffle receipt does not match approved delegation binding"
-                )
-            self._require_receipt_string_attribute(receipt, "external_receipt_id")
-            self._require_receipt_string_attribute(receipt, "coordination_target_id")
-            self._require_receipt_string_attribute(receipt, "ticket_reference_url")
-
-    def _finalized_execution_provenance(
-        self,
-        *,
-        execution: ActionExecutionRecord,
-        receipt: object,
-        execution_surface_id: str,
-    ) -> dict[str, object]:
-        provenance = dict(execution.provenance)
-        provenance["adapter"] = self._require_receipt_string_attribute(
-            receipt,
-            "adapter",
-        )
-        base_url = getattr(receipt, "base_url", "")
-        if isinstance(base_url, str) and base_url.strip() and base_url != "<set-me>":
-            provenance["adapter_base_url"] = base_url
-        if execution_surface_id == "shuffle":
-            provenance["downstream_binding"] = {
-                "approval_decision_id": self._require_receipt_string_attribute(
-                    receipt,
-                    "approval_decision_id",
-                ),
-                "delegation_id": self._require_receipt_string_attribute(
-                    receipt,
-                    "delegation_id",
-                ),
-                "payload_hash": self._require_receipt_string_attribute(
-                    receipt,
-                    "payload_hash",
-                ),
-            }
-            if execution.approved_payload.get("action_type") == "create_tracking_ticket":
-                provenance["downstream_binding"].update(
-                    {
-                        "coordination_reference_id": self._require_receipt_string_attribute(
-                            receipt,
-                            "coordination_reference_id",
-                        ),
-                        "coordination_target_type": self._require_receipt_string_attribute(
-                            receipt,
-                            "coordination_target_type",
-                        ),
-                        "external_receipt_id": self._require_receipt_string_attribute(
-                            receipt,
-                            "external_receipt_id",
-                        ),
-                        "coordination_target_id": self._require_receipt_string_attribute(
-                            receipt,
-                            "coordination_target_id",
-                        ),
-                        "ticket_reference_url": self._require_receipt_string_attribute(
-                            receipt,
-                            "ticket_reference_url",
-                        ),
-                    }
-                )
-        return provenance
-
-    @staticmethod
-    def _require_receipt_string_attribute(receipt: object, field_name: str) -> str:
-        value = getattr(receipt, field_name, None)
-        if not isinstance(value, str) or value.strip() == "":
-            raise ValueError(
-                f"adapter receipt missing required '{field_name}' attribute"
-            )
-        return value
-
-    def _mark_dispatch_failure(
-        self,
-        *,
-        execution: ActionExecutionRecord,
-        error: Exception,
-        receipt: object | None,
-    ) -> None:
-        failure_provenance = dict(execution.provenance)
-        dispatch_failure = {
-            "error": str(error),
-            "error_type": type(error).__name__,
-        }
-        if receipt is not None:
-            adapter = getattr(receipt, "adapter", None)
-            if isinstance(adapter, str) and adapter.strip():
-                failure_provenance["adapter"] = adapter
-            base_url = getattr(receipt, "base_url", None)
-            if isinstance(base_url, str) and base_url.strip() and base_url != "<set-me>":
-                failure_provenance["adapter_base_url"] = base_url
-            observed_surface_type = getattr(receipt, "execution_surface_type", None)
-            observed_surface_id = getattr(receipt, "execution_surface_id", None)
-            if isinstance(observed_surface_type, str) and observed_surface_type.strip():
-                dispatch_failure["observed_execution_surface_type"] = (
-                    observed_surface_type
-                )
-            if isinstance(observed_surface_id, str) and observed_surface_id.strip():
-                dispatch_failure["observed_execution_surface_id"] = observed_surface_id
-
-        failure_provenance["dispatch_failure"] = dispatch_failure
-        failure_at = datetime.now(timezone.utc)
-        with self._service._store.transaction():
-            current_execution = self._service._store.get(
-                ActionExecutionRecord,
-                execution.action_execution_id,
-            )
-            if current_execution is None or current_execution.lifecycle_state != "dispatching":
-                return
-            self._service.persist_record(
-                replace(
-                    current_execution,
-                    provenance=failure_provenance,
-                    lifecycle_state="failed",
-                ),
-                transitioned_at=failure_at,
-            )
 
     @staticmethod
     def _action_execution_lifecycle_from_status(
         status: object,
         current_lifecycle_state: str,
     ) -> str:
-        if not isinstance(status, str):
-            return current_lifecycle_state
-
-        normalized_status = status.strip().lower()
-        observed_state = {
-            "queued": "queued",
-            "pending": "queued",
-            "running": "running",
-            "in_progress": "running",
-            "success": "succeeded",
-            "succeeded": "succeeded",
-            "completed": "succeeded",
-            "failed": "failed",
-            "error": "failed",
-            "canceled": "canceled",
-            "cancelled": "canceled",
-        }.get(normalized_status)
-        if observed_state is None:
-            return current_lifecycle_state
-
-        normalized_current_lifecycle_state = current_lifecycle_state.strip().lower()
-        if normalized_current_lifecycle_state in {"succeeded", "failed", "canceled"}:
-            return current_lifecycle_state
-
-        lifecycle_rank = {
-            "dispatching": 0,
-            "queued": 1,
-            "running": 2,
-            "succeeded": 3,
-            "failed": 3,
-            "canceled": 3,
-        }
-        if lifecycle_rank[observed_state] >= lifecycle_rank.get(
-            normalized_current_lifecycle_state,
-            -1,
-        ):
-            return observed_state
-        return current_lifecycle_state
-
-    def _build_action_execution_reconciliation_key(
-        self,
-        *,
-        action_request: ActionRequestRecord,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        authoritative_execution: ActionExecutionRecord | None,
-    ) -> str:
-        components = [action_request.action_request_id]
-        if action_request.approval_decision_id is not None:
-            components.append(action_request.approval_decision_id)
-        if authoritative_execution is not None:
-            components.append(authoritative_execution.delegation_id)
-        components.extend(
-            (
-                execution_surface_type,
-                execution_surface_id,
-                action_request.idempotency_key,
-            )
+        return ActionExecutionReconciliationCoordinator._action_execution_lifecycle_from_status(
+            status,
+            current_lifecycle_state,
         )
-        return ":".join(components)
-
-    def _find_authoritative_action_execution(
-        self,
-        *,
-        action_request_id: str,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        execution_run_id: str | None,
-        idempotency_key: str,
-    ) -> ActionExecutionRecord | None:
-        matches = [
-            execution
-            for execution in self._service._store.list(ActionExecutionRecord)
-            if (
-                execution.action_request_id == action_request_id
-                and execution.execution_surface_type == execution_surface_type
-                and execution.execution_surface_id == execution_surface_id
-                and execution.idempotency_key == idempotency_key
-            )
-        ]
-        if execution_run_id is not None:
-            for execution in matches:
-                if execution.execution_run_id == execution_run_id:
-                    return execution
-        return matches[0] if matches else None
-
-    def _normalize_observed_executions(
-        self,
-        observed_executions: tuple[Mapping[str, object], ...],
-        *,
-        require_binding_identifiers: bool = False,
-        require_coordination_receipt_identifiers: bool = False,
-    ) -> tuple[dict[str, object], ...]:
-        normalized: list[dict[str, object]] = []
-        for execution in observed_executions:
-            execution_run_id = execution.get("execution_run_id")
-            execution_surface_id = execution.get("execution_surface_id")
-            idempotency_key = execution.get("idempotency_key")
-            observed_at = execution.get("observed_at")
-            approval_decision_id = execution.get("approval_decision_id")
-            delegation_id = execution.get("delegation_id")
-            payload_hash = execution.get("payload_hash")
-            coordination_reference_id = execution.get("coordination_reference_id")
-            coordination_target_type = execution.get("coordination_target_type")
-            external_receipt_id = execution.get("external_receipt_id")
-            coordination_target_id = execution.get("coordination_target_id")
-            ticket_reference_url = execution.get("ticket_reference_url")
-            if not isinstance(execution_run_id, str):
-                raise ValueError("observed execution must include string execution_run_id")
-            if not isinstance(execution_surface_id, str):
-                raise ValueError(
-                    "observed execution must include string execution_surface_id"
-                )
-            if not isinstance(idempotency_key, str):
-                raise ValueError("observed execution must include string idempotency_key")
-            if not isinstance(observed_at, datetime):
-                raise ValueError("observed execution must include datetime observed_at")
-            observed_at = self._service._require_aware_datetime(observed_at, "observed_at")
-            if require_binding_identifiers:
-                if not isinstance(approval_decision_id, str):
-                    raise ValueError(
-                        "observed execution must include string approval_decision_id"
-                    )
-                if not isinstance(delegation_id, str):
-                    raise ValueError("observed execution must include string delegation_id")
-                if not isinstance(payload_hash, str):
-                    raise ValueError("observed execution must include string payload_hash")
-            if require_coordination_receipt_identifiers:
-                if (
-                    not isinstance(coordination_reference_id, str)
-                    or not coordination_reference_id.strip()
-                ):
-                    raise ValueError(
-                        "observed execution must include string coordination_reference_id"
-                    )
-                if (
-                    not isinstance(coordination_target_type, str)
-                    or not coordination_target_type.strip()
-                ):
-                    raise ValueError(
-                        "observed execution must include string coordination_target_type"
-                    )
-                if (
-                    not isinstance(external_receipt_id, str)
-                    or not external_receipt_id.strip()
-                ):
-                    raise ValueError(
-                        "observed execution must include string external_receipt_id"
-                    )
-                if (
-                    not isinstance(coordination_target_id, str)
-                    or not coordination_target_id.strip()
-                ):
-                    raise ValueError(
-                        "observed execution must include string coordination_target_id"
-                    )
-                if (
-                    not isinstance(ticket_reference_url, str)
-                    or not ticket_reference_url.strip()
-                ):
-                    raise ValueError(
-                        "observed execution must include string ticket_reference_url"
-                    )
-            normalized.append(
-                {
-                    "execution_run_id": execution_run_id,
-                    "execution_surface_id": execution_surface_id,
-                    "idempotency_key": idempotency_key,
-                    "observed_at": observed_at,
-                    "approval_decision_id": approval_decision_id,
-                    "delegation_id": delegation_id,
-                    "payload_hash": payload_hash,
-                    "coordination_reference_id": coordination_reference_id,
-                    "coordination_target_type": coordination_target_type,
-                    "external_receipt_id": external_receipt_id,
-                    "coordination_target_id": coordination_target_id,
-                    "ticket_reference_url": ticket_reference_url,
-                    "status": execution.get("status"),
-                }
-            )
-
-        normalized.sort(key=lambda execution: execution["observed_at"])
-        return tuple(normalized)
-
-    def _load_approved_delegation_context(
-        self,
-        *,
-        action_request_id: str,
-        approved_payload: Mapping[str, object],
-        delegated_at: datetime,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        invalid_execution_surface_type_message: str,
-        invalid_execution_surface_id_message: str,
-        delegation_label: str,
-    ) -> tuple[ActionRequestRecord, ApprovalDecisionRecord]:
-        action_request = self._service._store.get(ActionRequestRecord, action_request_id)
-        if action_request is None:
-            raise LookupError(f"Missing action request {action_request_id!r}")
-        if action_request.lifecycle_state != "approved":
-            raise ValueError(
-                f"Action request {action_request_id!r} is not approved "
-                f"(state={action_request.lifecycle_state!r})"
-            )
-        approval_decision_id = self._service._require_non_empty_string(
-            action_request.approval_decision_id,
-            "action_request.approval_decision_id",
-        )
-        approval_decision = self._service._store.get(
-            ApprovalDecisionRecord,
-            approval_decision_id,
-        )
-        if approval_decision is None:
-            raise LookupError(
-                f"Missing approval decision {approval_decision_id!r} for action request "
-                f"{action_request_id!r}"
-            )
-        if approval_decision.lifecycle_state != "approved":
-            raise ValueError(
-                f"Approval decision {approval_decision_id!r} is not approved "
-                f"(state={approval_decision.lifecycle_state!r})"
-            )
-        if approval_decision.payload_hash != action_request.payload_hash:
-            raise ValueError(
-                "approval decision payload_hash does not match action request payload_hash"
-            )
-        policy_evaluation = action_request.policy_evaluation
-        if policy_evaluation.get("execution_surface_type") != execution_surface_type:
-            raise ValueError(invalid_execution_surface_type_message)
-        if policy_evaluation.get("execution_surface_id") != execution_surface_id:
-            raise ValueError(invalid_execution_surface_id_message)
-        self._require_exact_approved_payload_binding(
-            action_request=action_request,
-            approval_decision=approval_decision,
-            approved_payload=approved_payload,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-        )
-        self._require_exact_approved_expiry_binding(
-            action_request=action_request,
-            approval_decision=approval_decision,
-            delegated_at=delegated_at,
-            delegation_label=delegation_label,
-        )
-        return action_request, approval_decision
-
-    def _require_exact_approved_expiry_binding(
-        self,
-        *,
-        action_request: ActionRequestRecord,
-        approval_decision: ApprovalDecisionRecord,
-        delegated_at: datetime,
-        delegation_label: str,
-    ) -> None:
-        if approval_decision.approved_expires_at != action_request.expires_at:
-            raise ValueError("approved expiry window does not match action request expiry")
-        if (
-            approval_decision.approved_expires_at is not None
-            and delegated_at > approval_decision.approved_expires_at
-        ):
-            raise ValueError(
-                f"Action request {action_request.action_request_id!r} expired before {delegation_label} delegation"
-            )
-
-    def _require_exact_approved_payload_binding(
-        self,
-        *,
-        action_request: ActionRequestRecord,
-        approval_decision: ApprovalDecisionRecord,
-        approved_payload: Mapping[str, object],
-        execution_surface_type: str,
-        execution_surface_id: str,
-    ) -> None:
-        if approval_decision.action_request_id != action_request.action_request_id:
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
-        if approval_decision.target_snapshot != action_request.target_scope:
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
-
-        payload_hash = _approved_payload_binding_hash(
-            target_scope=action_request.target_scope,
-            approved_payload=approved_payload,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-        )
-        if (
-            payload_hash != action_request.payload_hash
-            or payload_hash != approval_decision.payload_hash
-        ):
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
