+++ control-plane/tests/test_execution_coordinator_boundary.py
@@ -12,6 +12,7 @@ if str(CONTROL_PLANE_ROOT) not in sys.path:
     sys.path.insert(0, str(CONTROL_PLANE_ROOT))
 
 from aegisops_control_plane.config import RuntimeConfig
+from aegisops_control_plane.execution_coordinator import ExecutionCoordinator
 from aegisops_control_plane.service import AegisOpsControlPlaneService
 from postgres_test_support import make_store
 
@@ -144,6 +145,138 @@ class ExecutionCoordinatorBoundaryTests(unittest.TestCase):
 
         self.assertIs(result, expected)
         coordinator.reconcile_action_execution.assert_called_once_with(
+            action_request_id="action-request-001",
+            execution_surface_type="automation_substrate",
+            execution_surface_id="shuffle",
+            observed_executions=observed_executions,
+            compared_at=compared_at,
+            stale_after=stale_after,
+        )
+
+    def test_execution_coordinator_composes_dedicated_internal_collaborators(self) -> None:
+        coordinator = ExecutionCoordinator(mock.sentinel.service)
+
+        self.assertTrue(
+            hasattr(coordinator, "_action_requests"),
+            "expected ExecutionCoordinator to compose a dedicated action request collaborator",
+        )
+        self.assertTrue(
+            hasattr(coordinator, "_delegation"),
+            "expected ExecutionCoordinator to compose a dedicated delegation collaborator",
+        )
+        self.assertTrue(
+            hasattr(coordinator, "_reconciliation"),
+            "expected ExecutionCoordinator to compose a dedicated reconciliation collaborator",
+        )
+
+    def test_execution_coordinator_routes_public_entrypoints_through_internal_collaborators(
+        self,
+    ) -> None:
+        coordinator = ExecutionCoordinator(mock.sentinel.service)
+        action_request_result = object()
+        shuffle_result = object()
+        isolated_result = object()
+        reconciliation_result = object()
+        coordinator._action_requests = mock.Mock()
+        coordinator._delegation = mock.Mock()
+        coordinator._reconciliation = mock.Mock()
+        coordinator._action_requests.create_reviewed_action_request_from_advisory.return_value = (
+            action_request_result
+        )
+        coordinator._delegation.delegate_approved_action_to_shuffle.return_value = (
+            shuffle_result
+        )
+        coordinator._delegation.delegate_approved_action_to_isolated_executor.return_value = (
+            isolated_result
+        )
+        coordinator._reconciliation.reconcile_action_execution.return_value = (
+            reconciliation_result
+        )
+        expires_at = datetime(2026, 4, 15, 0, 0, tzinfo=timezone.utc)
+        delegated_at = datetime(2026, 4, 15, 1, 0, tzinfo=timezone.utc)
+        compared_at = datetime(2026, 4, 15, 2, 0, tzinfo=timezone.utc)
+        stale_after = datetime(2026, 4, 15, 3, 0, tzinfo=timezone.utc)
+        observed_executions = (
+            {
+                "execution_run_id": "shuffle-run-001",
+                "execution_surface_id": "shuffle",
+                "idempotency_key": "idem-001",
+                "approval_decision_id": "approval-001",
+                "delegation_id": "delegation-001",
+                "payload_hash": "payload-001",
+                "observed_at": compared_at,
+            },
+        )
+
+        self.assertIs(
+            coordinator.create_reviewed_action_request_from_advisory(
+                record_family="recommendation",
+                record_id="recommendation-001",
+                requester_identity="analyst-001",
+                recipient_identity="owner-001",
+                message_intent="notify",
+                escalation_reason="bounded reason",
+                expires_at=expires_at,
+            ),
+            action_request_result,
+        )
+        self.assertIs(
+            coordinator.delegate_approved_action_to_shuffle(
+                action_request_id="action-request-001",
+                approved_payload={"action": "notify"},
+                delegated_at=delegated_at,
+                delegation_issuer="control-plane-service",
+                evidence_ids=("evidence-001",),
+            ),
+            shuffle_result,
+        )
+        self.assertIs(
+            coordinator.delegate_approved_action_to_isolated_executor(
+                action_request_id="action-request-002",
+                approved_payload={"action": "notify"},
+                delegated_at=delegated_at,
+                delegation_issuer="control-plane-service",
+                evidence_ids=("evidence-002",),
+            ),
+            isolated_result,
+        )
+        self.assertIs(
+            coordinator.reconcile_action_execution(
+                action_request_id="action-request-001",
+                execution_surface_type="automation_substrate",
+                execution_surface_id="shuffle",
+                observed_executions=observed_executions,
+                compared_at=compared_at,
+                stale_after=stale_after,
+            ),
+            reconciliation_result,
+        )
+
+        coordinator._action_requests.create_reviewed_action_request_from_advisory.assert_called_once_with(
+            record_family="recommendation",
+            record_id="recommendation-001",
+            requester_identity="analyst-001",
+            recipient_identity="owner-001",
+            message_intent="notify",
+            escalation_reason="bounded reason",
+            expires_at=expires_at,
+            action_request_id=None,
+        )
+        coordinator._delegation.delegate_approved_action_to_shuffle.assert_called_once_with(
+            action_request_id="action-request-001",
+            approved_payload={"action": "notify"},
+            delegated_at=delegated_at,
+            delegation_issuer="control-plane-service",
+            evidence_ids=("evidence-001",),
+        )
+        coordinator._delegation.delegate_approved_action_to_isolated_executor.assert_called_once_with(
+            action_request_id="action-request-002",
+            approved_payload={"action": "notify"},
+            delegated_at=delegated_at,
+            delegation_issuer="control-plane-service",
+            evidence_ids=("evidence-002",),
+        )
+        coordinator._reconciliation.reconcile_action_execution.assert_called_once_with(
             action_request_id="action-request-001",
             execution_surface_type="automation_substrate",
             execution_surface_id="shuffle",
